ref: separate cluster-agent services into different modules (#516)
