+++ crates/cluster_agent/src/main.rs
@@ -1,18 +1,19 @@
-use crate::services::ClusterAgent;
 use tonic::transport::Server;
 use types::cluster_agent::FILE_DESCRIPTOR_SET;
-use types::cluster_agent::log_metadata_service_server;
-use types::cluster_agent::log_records_service_server;
-mod services;
+use types::cluster_agent::log_metadata_service_server::LogMetadataServiceServer;
+use types::cluster_agent::log_records_service_server::LogRecordsServiceServer;
+mod logmetadata;
+mod logrecords;
+
+use logmetadata::LogMetadata;
+use logrecords::LogRecords;
 
 #[tokio::main]
 async fn main() -> eyre::Result<()> {
-    let server = ClusterAgent {};
-
     let (agent_health_reporter, agent_health_service) = tonic_health::server::health_reporter();
 
     agent_health_reporter
-        .set_serving::<log_metadata_service_server::LogMetadataServiceServer<ClusterAgent>>()
+        .set_serving::<LogMetadataServiceServer<LogMetadata>>()
         .await;
 
     let reflection_service = tonic_reflection::server::Builder::configure()
@@ -22,12 +23,8 @@ async fn main() -> eyre::Result<()> {
     Server::builder()
         .add_service(agent_health_service)
         .add_service(reflection_service)
-        .add_service(log_metadata_service_server::LogMetadataServiceServer::new(
-            server.clone(),
-        ))
-        .add_service(log_records_service_server::LogRecordsServiceServer::new(
-            server,
-        ))
+        .add_service(LogMetadataServiceServer::new(LogMetadata {}))
+        .add_service(LogRecordsServiceServer::new(LogRecords {}))
         .serve("[::1]:50051".parse()?)
         .await
         .unwrap();
